requirements.txt
isoweek>=1.3
tqdm>=4.19
bcolz>=1.1
kaggle_data>=0.4
opencv_python>=3.4
scikit_image>=0.13
setuptools>=38.4
numpy>=1.14
matplotlib>=2.1
Pillow>=5.0
scikit_learn>=0.19
tensorboardX>=1.1
typing>=3.6
PyYAML>=3.12
fuzzywuzzy>=0.16
python-Levenshtein
torch>=0.4.0
torchvision
tensorflow-gpu
